fixes bugs to render images & views from images array
