+++ src/images.js
@@ -1,19 +1,6 @@
 
-const images = [
-  {
-    id: new Date(),
-    title: 'Cute Bunny',
-    description: 'Isn\'t it fuzzy-wuzzy cutest thing you\'ve ever seen?',
-    url: 'http://f.cl.ly/items/3g3J1G0w122M360w380O/3726490195_f7cc75d377_o.jpg'   
-  },
-  {
-    id: new Date(),
-    title: 'Snow Flake',
-    description: 'White Rabbit',
-    url: 'http://im5.leaderhero.com/wallpaper/20140328/12835c9a-f.jpg'
-  },
-];
 
-export default function ImagesList() {
-  return images;
-}
+
+// export default function ImagesList() {
+//   return images;
+// }
